+++ ServerDevcommands/Features/AccessPrivateChests.cs
@@ -1,10 +1,10 @@
 using HarmonyLib;
 namespace ServerDevcommands;
 [HarmonyPatch(typeof(Container), nameof(Container.CheckAccess))]
-public class AccessPrivateChests {
-  static void Postfix(ref bool __result) {
-    __result |= Settings.AccessPrivateChests;
-  }
+public class AccessPrivateChests
+{
+  static bool Postfix(bool result) => result || Settings.AccessPrivateChests;
 }
+