+++ ServerDevcommands/Features/AutomaticPickUp.cs
@@ -1,8 +1,7 @@
 using HarmonyLib;
 namespace ServerDevcommands;
 [HarmonyPatch(typeof(Player), nameof(Player.Awake))]
-public class AutomaticPickUp {
-  static void Postfix(Player __instance) {
-    __instance.m_enableAutoPickup = Settings.AutomaticItemPickUp;
-  }
+public class AutomaticPickUp
+{
+  static void Postfix() => Player.m_enableAutoPickup = Settings.AutomaticItemPickUp;
 }